debugging in preparation for handlebars html that will hold data
